Split semantic_index_basic_test.cpp into focused test files
- Create semantic_index_test.cpp for core functionality (19 assertions)
- Create document_symbol_builder_test.cpp for document symbols (26 assertions)
- Create generate_block_test.cpp for SystemVerilog generate constructs (28 assertions)
- Create type_handling_test.cpp for enum/struct/typedef handling (30 assertions)
- Update BUILD.bazel with new test targets
- Remove original 875-line semantic_index_basic_test.cpp
- Add test size guidelines to CLAUDE.md
