+++ test/slangd/BUILD.bazel
@@ -70,20 +70,6 @@ cc_test(
     ],
 )
 
-cc_test(
-    name = "semantic_index_basic_test",
-    timeout = "short",
-    srcs = [
-        "semantic/semantic_index_basic_test.cpp",
-        "semantic/test_fixtures.hpp",
-    ],
-    deps = [
-        "//:slangd_core",
-        "@catch2",
-        "@fmt",
-        "@slang",
-    ],
-)
 
 cc_test(
     name = "semantic_index_patterns_test",
@@ -125,6 +111,66 @@ cc_test(
         ":simple_fixture",
         "//:slangd_core",
         "@catch2",
+        "@slang",
+    ],
+)
+
+cc_test(
+    name = "semantic_index_test",
+    timeout = "short",
+    srcs = [
+        "semantic/semantic_index_test.cpp",
+        "semantic/test_fixtures.hpp",
+    ],
+    deps = [
+        "//:slangd_core",
+        "@catch2",
+        "@fmt",
+        "@slang",
+    ],
+)
+
+cc_test(
+    name = "document_symbol_builder_test",
+    timeout = "short",
+    srcs = [
+        "semantic/document_symbol_builder_test.cpp",
+        "semantic/test_fixtures.hpp",
+    ],
+    deps = [
+        "//:slangd_core",
+        "@catch2",
+        "@fmt",
+        "@slang",
+    ],
+)
+
+cc_test(
+    name = "generate_block_test",
+    timeout = "short",
+    srcs = [
+        "semantic/generate_block_test.cpp",
+        "semantic/test_fixtures.hpp",
+    ],
+    deps = [
+        "//:slangd_core",
+        "@catch2",
+        "@fmt",
+        "@slang",
+    ],
+)
+
+cc_test(
+    name = "type_handling_test",
+    timeout = "short",
+    srcs = [
+        "semantic/type_handling_test.cpp",
+        "semantic/test_fixtures.hpp",
+    ],
+    deps = [
+        "//:slangd_core",
+        "@catch2",
+        "@fmt",
         "@slang",
     ],
 )
